refactor(module): extract `hastChildrenTreeParser` to utils
